// app/[locale]/(protected)/art/components/ArtworkForm/types.ts
export type FormFields = {
    register: UseFormRegister<any>
    control: Control<any>
    errors: FieldErrors
    setValue: UseFormSetValue<any>
    getValues: UseFormGetValues<any>
    isDirty?: boolean
    isSubmitted?: boolean
    onTitleChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
    onPricingOptionsChange?: (value: string) => void
    slug?: string
    title?: string
    isFormReadOnly?: boolean
    mediums?: ArtworkMedium[]
    styles?: ArtworkStyle[]
    techniques?: ArtworkTechnique[]
    themes?: ArtworkTheme[]
}
// ...
export interface PricingSectionProps extends FormFields {
    hasPhysicalOnly: boolean
    hasNftOnly: boolean
    hasNftPlusPhysical: boolean
    onPricingOptionChange: (option: 'hasPhysicalOnly' | 'hasNftOnly' | 'hasNftPlusPhysical', checked: boolean) => void
    isEditMode?: boolean
    physicalItemStatus?: string
    nftItemStatus?: string
    isFormReadOnly?: boolean
    isPhysicalOnly?: boolean
}